refactor: extract managed drive package
Move managed wake candidates, ledgers, render client, and Codex appserver turn driver into a dedicated drive package while keeping driver as a facade for existing callers.

Validation: go test ./harness/internal/drive ./harness/internal/driver ./harness/cmd/mnemond ./harness/cmd/mnemon-multica-runtime ./harness/cmd/mnemon-acceptance
